+++ packages/react/src/components/WhatsNew/__stories__/WhatsNew.mdx
@@ -4,15 +4,17 @@ import * as WhatsNewStories from './WhatsNew.stories.js';
 import * as TocStories from './Toc.stories';
 import * as ViewStackStories from './ViewStack.stories';
 import * as BubbleStories from './Bubble.stories';
-import { Toc } from '../components/Toc';
-import { TocItem } from '../components/TocItem';
-import { TocList } from '../components/TocList';
-import { TocSections } from '../components/TocSections';
-import { TocSection } from '../components/TocSection';
-import { Bubble } from '../components/Bubble';
-import { BubbleHeader } from '../components/BubbleHeader';
-import { ViewStack } from '../components/ViewStack';
-import { View } from '../components/View';
+import {
+  Toc,
+  TocList,
+  TocItem,
+  TocSection,
+  TocSections,
+  View,
+  ViewStack,
+  Bubble,
+  BubbleHeader,
+} from '../index';
 
 <Meta
   isTemplate
