+++ packages/react/src/components/WhatsNew/__tests__/Bubble.test.js
@@ -11,7 +11,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import React from 'react';
 import '@testing-library/jest-dom';
 import { Bubble } from '../components/Bubble';
-import { BubbleHeader } from '../components/BubbleHeader';
+import { BubbleHeader } from '../components/Bubble/BubbleHeader';
 
 jest.mock('./whats-new.scss', () => ({}));
