+++ packages/react/src/components/WhatsNew/__tests__/ViewStack.test.js
@@ -10,8 +10,7 @@
 import { render, act } from '@testing-library/react';
 import React from 'react';
 import '@testing-library/jest-dom';
-import { ViewStack } from '../components/ViewStack';
-import { View } from '../components/View';
+import { ViewStack, View } from '../components/ViewStack';
 
 jest.mock('./whats-new.scss', () => ({}));
 describe('ViewStack', () => {
