+++ packages/react/src/components/WhatsNew/components/Toc/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export { Toc } from './Toc';
+export { TocItem } from './TocItem';
+export { TocList } from './TocList';
+export { TocSections } from './TocSections';
+export { TocSection } from './TocSection';
