+++ packages/react/src/components/WhatsNew/components/ViewStack/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+export { ViewStack } from './ViewStack';
+export { View } from './View';
