+++ packages/react/src/components/WhatsNew/index.ts
@@ -6,12 +6,12 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-export { Bubble } from './components/Bubble.js';
-export { BubbleHeader } from './components/BubbleHeader.js';
-export { Toc } from './components/Toc.js';
-export { TocItem } from './components/TocItem.js';
-export { TocList } from './components/TocList.js';
-export { TocSection } from './components/TocSection.js';
-export { TocSections } from './components/TocSections.js';
-export { View } from './components/View.js';
-export { ViewStack } from './components/ViewStack.js';
+export { Bubble, BubbleHeader } from './components/Bubble';
+export {
+  Toc,
+  TocItem,
+  TocList,
+  TocSection,
+  TocSections,
+} from './components/Toc';
+export { ViewStack, View } from './components/ViewStack';
